+++ packages/sdk/src/erc20/index.ts
@@ -1,3 +1,4 @@
+export { erc20abi } from './abi/erc20abi.js';
 export { AbstractLidoSDKErc20 } from './erc20.js';
 export { LidoSDKstETH } from './steth.js';
 export { LidoSDKwstETH } from './wsteth.js';