+++ packages/sdk/tests/global-setup.cjs
@@ -2,49 +2,64 @@ const path = require('path');
 const dotenv = require('dotenv');
 const ganache = require('ganache');
 
-module.exports = async function () {
-  dotenv.config({
-    path: path.resolve(process.cwd(), '.env'),
-  });
-
-  const rpcUrl = process.env.TEST_RPC_URL;
-  const chainId = Number(process.env.TEST_CHAIN_ID);
-
+const setupGanacheProvider = async (chainId, rpcUrl) => {
   const ganacheProvider = ganache.provider({
     fork: { url: rpcUrl },
     logging: { quiet: true },
     chain: { chainId, asyncRequestProcessing: true },
   });
-
-  console.debug('\nInitializing ganache provider...');
+  console.debug(`\n[${chainId}]Initializing ganache provider...`);
   await ganacheProvider.initialize();
-  console.debug('Initialized ganache provider OK');
+  console.debug(`[${chainId}]Initialized ganache provider OK`);
 
-  console.debug('Testing direct RPC provider...');
+  console.debug(`[${chainId}]Testing direct RPC provider...`);
   const { result } = await fetch(rpcUrl, {
     method: 'POST',
     body: JSON.stringify({
       method: 'eth_chainId',
       params: [],
+      id: 1,
+      jsonrpc: '2.0',
     }),
     headers: {
       'Content-Type': 'application/json',
     },
   }).then((response) => response.json());
-  if (Number(result) !== chainId) {
-    throw new Error(`Invalid direct RPC provider response: ${result}`);
+  if (parseInt(result, 16) !== chainId) {
+    throw new Error(
+      `[${chainId}]Invalid direct RPC provider response: ${result}`,
+    );
   }
-  console.debug('Direct RPC provider OK');
+  console.debug(`[${chainId}]Direct RPC provider OK`);
 
-  console.debug('Testing ganache fork RPC provider...');
+  console.debug(`[${chainId}]Testing ganache fork RPC provider...`);
   const testRequest = await ganacheProvider.request({
     method: 'eth_chainId',
     params: [],
   });
-  if (Number(testRequest) !== chainId) {
-    throw new Error(`Invalid ganache response: ${testRequest}`);
+  if (parseInt(testRequest, 16) !== chainId) {
+    throw new Error(`[${chainId}]Invalid ganache response: ${testRequest}`);
   }
-  console.debug('Ganache fork RPC provider OK');
+  console.debug(`[${chainId}]Ganache fork RPC provider OK`);
+  return ganacheProvider;
+};
 
-  globalThis.__ganache_provider__ = ganacheProvider;
+module.exports = async function () {
+  dotenv.config({
+    path: path.resolve(process.cwd(), '.env'),
+  });
+
+  // L1
+  const chainId = Number(process.env.TEST_CHAIN_ID);
+  const rpcUrl = process.env.TEST_RPC_URL;
+  globalThis.__ganache_provider__ = await setupGanacheProvider(chainId, rpcUrl);
+
+  // L2
+  const l2RpcUrl = process.env.TEST_L2_RPC_URL;
+  const l2ChainId = Number(process.env.TEST_L2_CHAIN_ID);
+
+  globalThis.__l2_ganache_provider__ = await setupGanacheProvider(
+    l2ChainId,
+    l2RpcUrl,
+  );
 };