+++ playground/components/layout/header/headerWallet.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
-import { CHAINS, getChainColor } from '@lido-sdk/constants';
+import { getChainColor } from '@lido-sdk/constants';
 
 import { useWeb3 } from 'reef-knot/web3-react';
 import { ThemeToggler } from '@lidofinance/lido-ui';
@@ -9,15 +9,27 @@ import WalletConnect from 'components/layout/header/walletConnect';
 
 import { HeaderWalletChainStyle } from './headerWalletStyles';
 
+import { useChains } from 'wagmi';
+
+const tryGetColor = (chainId: number) => {
+  try {
+    return getChainColor(chainId);
+  } catch {
+    return '#FFFFFF';
+  }
+};
+
 const HeaderWallet: FC = () => {
   const { active, chainId } = useWeb3();
-  const chainName = chainId && CHAINS[chainId];
+  const chains = useChains();
+
+  const currentChain = chains.find((chain) => chain.id === chainId);
 
   return (
     <>
       {chainId && (
-        <HeaderWalletChainStyle $color={getChainColor(chainId)}>
-          {chainName}
+        <HeaderWalletChainStyle $color={tryGetColor(chainId)}>
+          {currentChain?.name}
         </HeaderWalletChainStyle>
       )}
       {active ? <WalletButton /> : <WalletConnect size="sm" />}