+++ playground/providers/web3.tsx
@@ -24,6 +24,8 @@ import { useThemeToggle } from '@lidofinance/lido-ui';
 
 type ChainsList = [wagmiChains.Chain, ...wagmiChains.Chain[]];
 
+export const L2_CHAINS = [10, 11155420];
+
 const wagmiChainsArray = Object.values(wagmiChains) as any as ChainsList;
 
 const supportedChains = wagmiChainsArray.filter((chain) =>
@@ -63,6 +65,8 @@ const Web3Provider: FC<PropsWithChildren> = ({ children }) => {
       [CHAINS.Goerli]: getRpc(CHAINS.Goerli),
       [CHAINS.Holesky]: getRpc(CHAINS.Holesky),
       [CHAINS.Sepolia]: getRpc(CHAINS.Sepolia),
+      // OP sepolia
+      [11155420]: getRpc(11155420),
     };
   }, [customRpc]);